arch: make _start arch specific
For now we don't have anything useful for testing, so lets make our _start()
a arch specific function (boot.s) to make it easier to see the the ASM
output in QEMU (`make qemu`). But for now, with this commit, our _start only
waits for an event and loop back to the event wait in case something
happens.