drivers: add basic infra for device drivers
This commit adds a simple infrastructure to manage the drivers we plan to
have in Lapwing in the first stages of development. The idea is to keep
track of each driver by allowing each of them to register themselves within
a single manager. There should be only one drivers manager in the kernel
running at any given time, which should be declared in the static lifetime,
being the reason the reference to the drivers also be &'static.

Some locking mechanism is required or some way to let Rust compiler know the
manager will only be created and managed in single thread/core mode with
interrupts disabled.