bsp/mmio: add basic MMIO address access code
This commit brings the abstraction for MMIO address specific for RPi4B,
adding the ability to translate the address from one peripheric mode to
another (when necessary) and also setting the base address of MMIO registers
for later being used for hardware/protocol layers setup.

The base address is set on mmio::MemMap creation and the access happens
through the dereferencing operation (impl Deref trait).