klib/sync: make SingleThreadData inner data UnsafeCell
Instead of relying on Deref(Mut) traits for the inner data of
SingleThreadData we moved to the internal mutability concept with
UnsafeCell, which matches the definition of a static data, by which this
struct is mostly used. This help us to avoid mixing concepts and making
things too confusing.

In order to get Deref(Mut) to work properly we would have to manage custom
lifetimes for matching static data and advanced/hackish constructs to get
the compiler to accept dereferencing a "possibly" mutable static data (not
sure if it is even possible), thus using the internal mutability with
UnsafeCell fits nicely.